+++ app/assets/javascripts/application.js
@@ -3,6 +3,7 @@
 
 //= require compute_word_frequencies
 //= require datasets
+//= require named_entities
 //= require normalize_document_counts
 //= require plot_dates
 //= require search
@@ -11,5 +12,10 @@
 
 //= require jquery.mobile
 
-// Load up the Google Visualization API
+// Load up the Google Visualization and Maps APIs
+// FIXME: How does the user specify a Google Maps API key?
 google.load('visualization', '1.0', {'packages':['corechart','table']});
+google.load('maps', '3', {'other_params':'sensor=false'});
+google.setOnLoadCallback(function() {
+  google.maps.visualRefresh = true;
+});